+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Dash board updates, including Files page, in https://github.com/punch-mission/punchpipe/pull/242
 * Misc pipeline optimizations and fixes in https://github.com/punch-mission/punchpipe/pull/243
 * Dashboard code moved out of cli.py in https://github.com/punch-mission/punchpipe/pull/245
+* Reorganize L1, and improve L2/Q incomplete mosaic scheduling in https://github.com/punch-mission/punchpipe/pull/246
 
 ## Version 0.0.12: August 6, 2025